Move initial read logic into generic file
