+++ vortex-file/src/generic.rs
@@ -1,17 +1,17 @@
 use std::ops::Range;
-use std::sync::Arc;
+use std::sync::{Arc, RwLock};
 
 use futures::{StreamExt, pin_mut};
+use vortex_array::aliases::hash_map::HashMap;
 use vortex_buffer::{Alignment, ByteBuffer, ByteBufferMut};
-use vortex_error::{VortexExpect, VortexResult};
+use vortex_error::{VortexExpect, VortexResult, vortex_err};
 use vortex_io::{Dispatch, InstrumentedReadAt, IoDispatcher, VortexReadAt};
-use vortex_layout::segments::{SegmentEvents, SegmentSource};
+use vortex_layout::segments::{SegmentEvents, SegmentId, SegmentSource};
 use vortex_metrics::VortexMetrics;
 
 use crate::driver::CoalescedDriver;
-use crate::open::ParseResult;
 use crate::segments::{
-    InitialReadSegmentCache, MokaSegmentCache, SegmentCache, SegmentCacheMetrics,
+    InitialReadSegmentCache, MokaSegmentCache, NoOpSegmentCache, SegmentCache, SegmentCacheMetrics,
     SegmentCacheSourceAdapter,
 };
 use crate::{
@@ -41,6 +41,23 @@ impl VortexOpenOptions<GenericVortexFile> {
             .with_initial_read_size(Self::INITIAL_READ_SIZE)
     }
 
+    /// Configure the initial read size for the Vortex file.
+    pub fn with_initial_read_size(mut self, initial_read_size: u64) -> Self {
+        self.options.initial_read_size = initial_read_size;
+        self
+    }
+
+    /// Configure a custom [`SegmentCache`].
+    pub fn with_segment_cache(mut self, segment_cache: Arc<dyn SegmentCache>) -> Self {
+        self.options.segment_cache = segment_cache;
+        self
+    }
+
+    /// Disable segment caching entirely.
+    pub fn without_segment_cache(self) -> Self {
+        self.with_segment_cache(Arc::new(NoOpSegmentCache))
+    }
+
     pub fn with_io_concurrency(mut self, io_concurrency: usize) -> Self {
         self.options.io_concurrency = io_concurrency;
         self
@@ -52,13 +69,13 @@ impl VortexOpenOptions<GenericVortexFile> {
         let footer = if let Some(footer) = self.footer {
             footer
         } else {
-            self.read_footer(&read).await?
+            self.read_footer(read.clone()).await?
         };
 
         let segment_cache = Arc::new(SegmentCacheMetrics::new(
             InitialReadSegmentCache {
-                initial: self.initial_read_segments,
-                fallback: self.segment_cache,
+                initial: self.options.initial_read_segments,
+                fallback: self.options.segment_cache,
             },
             self.metrics.clone(),
         ));
@@ -67,7 +84,8 @@ impl VortexOpenOptions<GenericVortexFile> {
             read,
             segment_map: footer.segment_map().clone(),
             segment_cache,
-            options: self.options,
+            io_dispatcher: self.options.io_dispatcher,
+            io_concurrency: self.options.io_concurrency,
         });
 
         Ok(VortexFile {
@@ -77,69 +95,140 @@ impl VortexOpenOptions<GenericVortexFile> {
         })
     }
 
-    async fn read_footer<R: VortexReadAt + Send + Sync>(&self, read: &R) -> VortexResult<Footer> {
+    async fn read_footer<R: VortexReadAt + Send + Sync>(
+        &self,
+        read: Arc<R>,
+    ) -> VortexResult<Footer> {
         // Fetch the file size and perform the initial read.
         let file_size = match self.file_size {
-            None => self.dispatched_size(read).await?,
+            None => self.dispatched_size(read.clone()).await?,
             Some(file_size) => file_size,
         };
         let initial_read_size = self
+            .options
             .initial_read_size
             // Make sure we read enough to cover the postscript
             .max(MAX_FOOTER_SIZE as u64 + EOF_SIZE as u64)
             .min(file_size);
         let mut initial_offset = file_size - initial_read_size;
         let mut initial_read: ByteBuffer = self
-            .dispatched_read(read, initial_offset..file_size)
+            .dispatched_read(read.clone(), initial_offset..file_size)
             .await?;
 
         let postscript = self.parse_postscript(&initial_read)?;
 
-        loop {
-            match self.parse_footer(&initial_read).await? {
-                ParseResult::Ok(footer) => return Ok(footer),
-                ParseResult::NeedFrom(need_more_offset) => {
-                    // Do another range read.
-                    assert!(
-                        initial_offset > need_more_offset,
-                        "Requested fewer bytes than we passed?"
-                    );
-                    let mut need_more_buffer =
-                        ByteBufferMut::with_capacity((file_size - need_more_offset).try_into()?);
-                    need_more_buffer.extend_from_slice(
-                        &self
-                            .dispatched_read(read, need_more_offset..initial_offset)
-                            .await?,
-                    );
-                    need_more_buffer.extend_from_slice(&initial_read);
+        // If we haven't been provided a DType, we must read one from the file.
+        let dtype_segment = self.dtype.is_none().then(|| postscript.dtype.ok_or_else(|| vortex_err!("Vortex file doesn't embed a DType and one has not been provided to VortexOpenOptions"))).transpose()?;
 
-                    initial_read = need_more_buffer.into();
-                    initial_offset = need_more_offset;
-                }
-            }
+        // The other postscript segments are required, so now we figure out our the offset that
+        // contains all the required segments.
+        let mut read_more_offset = initial_offset;
+        if let Some(dtype_segment) = &dtype_segment {
+            read_more_offset = read_more_offset.min(dtype_segment.offset);
         }
+        if let Some(stats_segment) = &postscript.statistics {
+            read_more_offset = read_more_offset.min(stats_segment.offset);
+        }
+        read_more_offset = read_more_offset.min(postscript.layout.offset);
+
+        // Read more bytes if necessary.
+        if read_more_offset < initial_offset {
+            log::info!(
+                "Initial read from {} did not cover all footer segments, reading from {}",
+                initial_offset,
+                read_more_offset
+            );
+
+            let mut new_initial_read =
+                ByteBufferMut::with_capacity(usize::try_from(file_size - read_more_offset)?);
+            new_initial_read.extend_from_slice(
+                &self
+                    .dispatched_read(read, read_more_offset..initial_offset)
+                    .await?,
+            );
+            new_initial_read.extend_from_slice(&initial_read);
+
+            initial_offset = read_more_offset;
+            initial_read = new_initial_read.freeze();
+        }
+
+        // Now we read our initial segments.
+        let dtype = dtype_segment
+            .map(|segment| self.parse_dtype(initial_offset, &initial_read, &segment))
+            .transpose()?
+            .unwrap_or_else(|| self.dtype.clone().vortex_expect("DType was provided"));
+        let file_stats = postscript
+            .statistics
+            .map(|segment| self.parse_file_statistics(initial_offset, &initial_read, &segment))
+            .transpose()?;
+        let footer = self.parse_footer(
+            initial_offset,
+            &initial_read,
+            &postscript.layout,
+            dtype,
+            file_stats,
+        )?;
+
+        // If the initial read happened to cover any segments, then we can populate the
+        // segment cache
+        self.populate_initial_segments(initial_offset, &initial_read, &footer);
+
+        Ok(footer)
     }
 
     /// Dispatch a [`VortexReadAt::size`] request onto the configured I/O dispatcher.
-    async fn dispatched_size<R: VortexReadAt + Send>(&self, read: &R) -> VortexResult<u64> {
+    async fn dispatched_size<R: VortexReadAt + Send + Sync>(
+        &self,
+        read: Arc<R>,
+    ) -> VortexResult<u64> {
         Ok(self
             .options
             .io_dispatcher
-            .dispatch(|| async move { read.size().await })?
+            .dispatch(move || async move { read.size().await })?
             .await??)
     }
 
     /// Dispatch a read onto the configured I/O dispatcher.
-    async fn dispatched_read<R: VortexReadAt + Send>(
+    async fn dispatched_read<R: VortexReadAt + Send + Sync>(
         &self,
-        read: &R,
+        read: Arc<R>,
         range: Range<u64>,
     ) -> VortexResult<ByteBuffer> {
         Ok(self
             .options
             .io_dispatcher
-            .dispatch(|| async move { read.read_byte_range(range, Alignment::none()).await })?
+            .dispatch(move || async move { read.read_byte_range(range, Alignment::none()).await })?
             .await??)
+    }
+
+    /// Populate segments in the cache that were covered by the initial read.
+    fn populate_initial_segments(
+        &self,
+        initial_offset: u64,
+        initial_read: &ByteBuffer,
+        footer: &Footer,
+    ) {
+        let first_idx = footer
+            .segment_map()
+            .partition_point(|segment| segment.offset < initial_offset);
+
+        let mut initial_segments = self
+            .options
+            .initial_read_segments
+            .write()
+            .vortex_expect("poisoned lock");
+
+        for idx in first_idx..footer.segment_map().len() {
+            let segment = &footer.segment_map()[idx];
+            let segment_id =
+                SegmentId::from(u32::try_from(idx).vortex_expect("Invalid segment ID"));
+            let offset =
+                usize::try_from(segment.offset - initial_offset).vortex_expect("Invalid offset");
+            let buffer = initial_read
+                .slice(offset..offset + (segment.length as usize))
+                .aligned(segment.alignment);
+            initial_segments.insert(segment_id, buffer);
+        }
     }
 }
 
@@ -147,7 +236,8 @@ struct GenericVortexFileIo<R> {
     read: Arc<R>,
     segment_map: Arc<[SegmentSpec]>,
     segment_cache: Arc<dyn SegmentCache>,
-    options: GenericFileOptions,
+    io_dispatcher: IoDispatcher,
+    io_concurrency: usize,
 }
 
 impl<R: VortexReadAt + Send + Sync> SegmentSourceFactory for GenericVortexFileIo<R> {
@@ -171,9 +261,8 @@ impl<R: VortexReadAt + Send + Sync> SegmentSourceFactory for GenericVortexFileIo
         );
 
         // Spawn an I/O driver onto the dispatcher.
-        let io_concurrency = self.options.io_concurrency;
-        self.options
-            .io_dispatcher
+        let io_concurrency = self.io_concurrency;
+        self.io_dispatcher
             .dispatch(move || {
                 async move {
                     // Drive the segment event stream.
@@ -222,8 +311,10 @@ impl VortexOpenOptions<GenericVortexFile> {
     }
 }
 
-#[derive(Clone)]
 pub struct GenericFileOptions {
+    segment_cache: Arc<dyn SegmentCache>,
+    initial_read_size: u64,
+    initial_read_segments: RwLock<HashMap<SegmentId, ByteBuffer>>,
     /// The number of concurrent I/O requests to spawn.
     /// This should be smaller than execution concurrency for coalescing to occur.
     io_concurrency: usize,
@@ -234,6 +325,9 @@ pub struct GenericFileOptions {
 impl Default for GenericFileOptions {
     fn default() -> Self {
         Self {
+            segment_cache: Arc::new(NoOpSegmentCache),
+            initial_read_size: 0,
+            initial_read_segments: Default::default(),
             io_concurrency: 8,
             io_dispatcher: IoDispatcher::shared(),
         }
