+++ vortex-file/src/memory.rs
@@ -26,9 +26,30 @@ impl VortexOpenOptions<InMemoryVortexFile> {
     }
 
     /// Open an in-memory file contained in the provided buffer.
+    // TODO(ngates): we could add a lifetime to a VortexFile and then just take a &[u8] here?
     pub async fn open<B: Into<ByteBuffer>>(self, buffer: B) -> VortexResult<VortexFile> {
         let buffer = buffer.into();
-        let footer = self.parse_footer(&buffer).await?;
+
+        let postscript = self.parse_postscript(&buffer)?;
+
+        // If we haven't been provided a DType, we must read one from the file.
+        let dtype = self.dtype
+            .clone()
+            .map(Ok)
+            .unwrap_or_else(|| {
+                let dtype_segment = postscript
+                    .dtype
+                    .ok_or_else(|| vortex_err!("Vortex file doesn't embed a DType and one has not been provided to VortexOpenOptions"))?;
+                self.parse_dtype(0, &buffer, &dtype_segment)
+            })?;
+
+        let file_stats = postscript
+            .statistics
+            .map(|segment| self.parse_file_statistics(0, &buffer, &segment))
+            .transpose()?;
+
+        let footer = self.parse_footer(0, &buffer, &postscript.layout, dtype, file_stats)?;
+
         let segment_source_factory = Arc::new(InMemorySegmentReader {
             buffer,
             footer: footer.clone(),
