+++ vortex-file/src/open.rs
@@ -1,20 +1,15 @@
-use std::sync::{Arc, RwLock};
+use std::sync::Arc;
 
 use flatbuffers::root;
 use vortex_array::ArrayRegistry;
-use vortex_array::aliases::hash_map::HashMap;
-use vortex_buffer::{Alignment, ByteBuffer, ByteBufferMut};
 use vortex_dtype::DType;
-use vortex_error::{VortexError, VortexExpect, VortexResult, vortex_bail, vortex_err};
+use vortex_error::{VortexResult, vortex_bail, vortex_err};
 use vortex_flatbuffers::{FlatBuffer, ReadFlatBuffer, dtype as fbd};
-use vortex_io::VortexReadAt;
-use vortex_layout::segments::SegmentId;
 use vortex_layout::{LayoutRegistry, LayoutRegistryExt};
 use vortex_metrics::VortexMetrics;
 
 use crate::footer::{FileStatistics, Footer, Postscript, PostscriptSegment};
-use crate::segments::{NoOpSegmentCache, SegmentCache};
-use crate::{DEFAULT_REGISTRY, EOF_SIZE, MAGIC_BYTES, MAX_FOOTER_SIZE, VERSION};
+use crate::{DEFAULT_REGISTRY, EOF_SIZE, MAGIC_BYTES, VERSION};
 
 pub trait FileType: Sized {
     type Options;
@@ -35,9 +30,7 @@ pub struct VortexOpenOptions<F: FileType> {
     /// An optional, externally provided, file layout.
     // TODO(ngates): add an optional DType so we only read the layout segment.
     pub(crate) footer: Option<Footer>,
-    pub(crate) segment_cache: Arc<dyn SegmentCache>,
-    pub(crate) initial_read_size: u64,
-    pub(crate) initial_read_segments: RwLock<HashMap<SegmentId, ByteBuffer>>,
+    /// A metrics registry for the file.
     pub(crate) metrics: VortexMetrics,
 }
 
@@ -50,9 +43,6 @@ impl<F: FileType> VortexOpenOptions<F> {
             file_size: None,
             dtype: None,
             footer: None,
-            segment_cache: Arc::new(NoOpSegmentCache),
-            initial_read_size: 0,
-            initial_read_segments: Default::default(),
             metrics: VortexMetrics::default(),
         }
     }
@@ -98,23 +88,6 @@ impl<F: FileType> VortexOpenOptions<F> {
         self
     }
 
-    /// Configure the initial read size for the Vortex file.
-    pub fn with_initial_read_size(mut self, initial_read_size: u64) -> Self {
-        self.initial_read_size = initial_read_size;
-        self
-    }
-
-    /// Configure a custom [`SegmentCache`].
-    pub fn with_segment_cache(mut self, segment_cache: Arc<dyn SegmentCache>) -> Self {
-        self.segment_cache = segment_cache;
-        self
-    }
-
-    /// Disable segment caching entirely.
-    pub fn without_segment_cache(self) -> Self {
-        self.with_segment_cache(Arc::new(NoOpSegmentCache))
-    }
-
     /// Configure a custom [`VortexMetrics`].
     pub fn with_metrics(mut self, metrics: VortexMetrics) -> Self {
         self.metrics = metrics;
@@ -122,89 +95,7 @@ impl<F: FileType> VortexOpenOptions<F> {
     }
 }
 
-pub(crate) enum ParseResult<T> {
-    /// The initial read was successful.
-    Ok(T),
-    /// Parsing the value requires more bytes, specifically starting from
-    /// the given offset from the start of the file.
-    // TODO(ngates): we could measure this in usize from EOF if PostscriptSegments were defined
-    //  in negative offsets.
-    NeedFrom(u64),
-}
-
 impl<F: FileType> VortexOpenOptions<F> {
-    /// Parse the [`Footer`] out of the provided initial read buffer. The buffer must contain
-    /// the last N bytes of the file, where N is the size of the buffer itself. i.e. the buffer
-    /// must end with the `EOF` bytes.
-    ///
-    /// The function will ask for more bytes if necessary in order to fully read the footer.
-    ///
-    /// The alignment of the initial read buffer is not important.
-    pub(crate) fn parse_footer(&self, initial_read: &[u8]) -> VortexResult<ParseResult<Footer>> {
-        // We know the initial read _must_ contain at least the Postscript.
-        let postscript = self.parse_postscript(&initial_read)?;
-
-        // If we haven't been provided a DType, we must read one from the file.
-        let dtype_segment = self.dtype.is_none().then(|| postscript.dtype.ok_or_else(|| vortex_err!("Vortex file doesn't embed a DType and one has not been provided to VortexOpenOptions"))).transpose()?;
-
-        // The other postscript segments are required, so now we figure out our the offset that
-        // contains all the required segments.
-        let mut read_more_offset = initial_offset;
-        if let Some(dtype_segment) = &dtype_segment {
-            read_more_offset = read_more_offset.min(dtype_segment.offset);
-        }
-        if let Some(stats_segment) = &postscript.statistics {
-            read_more_offset = read_more_offset.min(stats_segment.offset);
-        }
-        read_more_offset = read_more_offset.min(postscript.layout.offset);
-
-        // Read more bytes if necessary.
-        if read_more_offset < initial_offset {
-            log::info!(
-                "Initial read from {} did not cover all footer segments, reading from {}",
-                initial_offset,
-                read_more_offset
-            );
-
-            let mut new_initial_read =
-                ByteBufferMut::with_capacity(usize::try_from(file_size - read_more_offset)?);
-            new_initial_read.extend_from_slice(
-                &read
-                    .read_byte_range(read_more_offset..initial_offset, Alignment::none())
-                    .await?,
-            );
-            new_initial_read.extend_from_slice(&initial_read);
-
-            initial_offset = read_more_offset;
-            initial_read = new_initial_read.freeze();
-        }
-
-        // Now we read our initial segments.
-        let dtype = dtype_segment
-            .map(|segment| self.parse_dtype(initial_offset, &initial_read, &segment))
-            .transpose()?
-            .unwrap_or_else(|| self.dtype.clone().vortex_expect("DType was provided"));
-        let file_stats = postscript
-            .statistics
-            .map(|segment| {
-                self.parse_flatbuffer::<FileStatistics>(initial_offset, &initial_read, &segment)
-            })
-            .transpose()?;
-        let footer = self.parse_file_layout(
-            initial_offset,
-            &initial_read,
-            &postscript.layout,
-            dtype,
-            file_stats,
-        )?;
-
-        // If the initial read happened to cover any segments, then we can populate the
-        // segment cache
-        self.populate_initial_segments(initial_offset, &initial_read, &footer);
-
-        Ok(footer)
-    }
-
     /// Parse the postscript from the initial read.
     pub(crate) fn parse_postscript(&self, initial_read: &[u8]) -> VortexResult<Postscript> {
         if initial_read.len() < EOF_SIZE {
@@ -236,10 +127,10 @@ impl<F: FileType> VortexOpenOptions<F> {
                 .map_err(|e| vortex_err!("Postscript size was not a u16 {e}"))?,
         ) as usize;
 
-        if initial_read.len() < eof_loc + ps_size {
+        if initial_read.len() < ps_size + EOF_SIZE {
             vortex_bail!(
                 "Initial read must be at least {} bytes to include the Postscript",
-                eof_loc + ps_size
+                ps_size + EOF_SIZE
             );
         }
 
@@ -247,46 +138,45 @@ impl<F: FileType> VortexOpenOptions<F> {
     }
 
     /// Parse the DType from the initial read.
-    fn parse_dtype(
+    pub(crate) fn parse_dtype(
         &self,
         initial_offset: u64,
-        initial_read: &ByteBuffer,
+        initial_read: &[u8],
         segment: &PostscriptSegment,
     ) -> VortexResult<DType> {
         let offset = usize::try_from(segment.offset - initial_offset)?;
         let sliced_buffer =
-            FlatBuffer::align_from(initial_read.slice(offset..offset + (segment.length as usize)));
+            FlatBuffer::copy_from(&initial_read[offset..offset + (segment.length as usize)]);
         let fbd_dtype = root::<fbd::DType>(&sliced_buffer)?;
 
         DType::try_from_view(fbd_dtype, sliced_buffer.clone())
     }
 
-    /// Parse a [`ReadFlatBuffer`] from the initial read buffer.
-    fn parse_flatbuffer<T: ReadFlatBuffer<Error = VortexError>>(
+    /// Parse the [`FileStatistics`] from the initial read buffer.
+    pub(crate) fn parse_file_statistics(
         &self,
         initial_offset: u64,
-        initial_read: &ByteBuffer,
+        initial_read: &[u8],
         segment: &PostscriptSegment,
-    ) -> VortexResult<T> {
+    ) -> VortexResult<FileStatistics> {
         let offset = usize::try_from(segment.offset - initial_offset)?;
         let sliced_buffer =
-            FlatBuffer::align_from(initial_read.slice(offset..offset + (segment.length as usize)));
-        T::read_flatbuffer_bytes(&sliced_buffer)
+            FlatBuffer::copy_from(&initial_read[offset..offset + (segment.length as usize)]);
+        FileStatistics::read_flatbuffer_bytes(&sliced_buffer)
     }
 
     /// Parse the rest of the footer from the initial read.
-    fn parse_file_layout(
+    pub(crate) fn parse_footer(
         &self,
         initial_offset: u64,
-        initial_read: &ByteBuffer,
+        initial_read: &[u8],
         layout_segment: &PostscriptSegment,
         dtype: DType,
         file_stats: Option<FileStatistics>,
     ) -> VortexResult<Footer> {
         let offset = usize::try_from(layout_segment.offset - initial_offset)?;
-        let bytes = FlatBuffer::align_from(
-            initial_read.slice(offset..offset + (layout_segment.length as usize)),
-        );
+        let bytes =
+            FlatBuffer::copy_from(&initial_read[offset..offset + (layout_segment.length as usize)]);
         Footer::from_flatbuffer(
             bytes,
             dtype,
@@ -294,34 +184,5 @@ impl<F: FileType> VortexOpenOptions<F> {
             &self.registry,
             &self.layout_registry,
         )
-    }
-
-    /// Populate segments in the cache that were covered by the initial read.
-    fn populate_initial_segments(
-        &self,
-        initial_offset: u64,
-        initial_read: &ByteBuffer,
-        footer: &Footer,
-    ) {
-        let first_idx = footer
-            .segment_map()
-            .partition_point(|segment| segment.offset < initial_offset);
-
-        let mut initial_segments = self
-            .initial_read_segments
-            .write()
-            .vortex_expect("poisoned lock");
-
-        for idx in first_idx..footer.segment_map().len() {
-            let segment = &footer.segment_map()[idx];
-            let segment_id =
-                SegmentId::from(u32::try_from(idx).vortex_expect("Invalid segment ID"));
-            let offset =
-                usize::try_from(segment.offset - initial_offset).vortex_expect("Invalid offset");
-            let buffer = initial_read
-                .slice(offset..offset + (segment.length as usize))
-                .aligned(segment.alignment);
-            initial_segments.insert(segment_id, buffer);
-        }
     }
 }
